+++ src/LivrEtec.GIB/LivrEtec.GIB.csproj
@@ -13,6 +13,7 @@
 		<PrivateAssets>all</PrivateAssets>
 		<IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
 	</PackageReference>
+	<PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="7.0.0" />
 </ItemGroup>
 <ItemGroup>
   <ProjectReference Include="..\LivrEtec\LivrEtec.csproj" />
